requirements.txt
aiogram==3.16.0
aiohttp==3.9.0
vk-api==11.9.9
python-dotenv==1.0.0